forgot to commit that one
git-svn-id: svn://svn.minisip.org/minisip/trunk@1132 63dae60f-c5d3-0310-92a3-fcbf00261c72

+++ minisip/soundcard/SoundSource.cxx
@@ -173,7 +173,7 @@ void BasicSoundSource::getSound(short *dest,
 #endif
                 if (plcProvider){
                         cerr << "PLC!"<< endl;
-                        short *b = plcProvider->get_plc_sound((int32_t)oFrames);
+                        short *b = plcProvider->get_plc_sound(oFrames);
                         memcpy(dest, b, oFrames);
                 }else{
